requirements.txt

langchain>=0.0.290
chromadb>=0.3.30
sentence-transformers>=2.2.2
requests>=2.31.0
openai>=0.27.0
tqdm>=4.65.0
python-dotenv>=1.0.0
uvicorn>=0.22.0
fastapi>=0.95.0
streamlit>=1.24.0
pytest>=7.4.0
